New game Engine
New features:

Text now has a new centre function - this is optional.
To use it, set the option variable to True when calling the function. This will take the X and Y coord and half them. I recommend when entering the coords of X and Y and set them as the width and height of your window so that it can find the centre of the screen.

New and improved .player.collisons() feature!
Requred variables are the player object, the list to search for collisions and the index in that list.
This will return the name of the .name of the object collided with the player. If it hasn't found an object to collide with, it returns None


Change logs:

.update.pygame_debug() - > .window.pygame_debug()
.update.list_debug() -> .window.list_debug()
.update.define() -> .window.define()
.update.window() ->.window.update()