New game engine.
Change logs:

.properties_object(file_location) -> .properties_object(loaded_texture):

Textures will now be accepted instead of file locations. This will decrease the time to create a game object: as it doesn't need to accessed from secondary storage.
This can help imrpove the performance of some programs that need to create a lot of similar objects with the same texture.
However, entered a file location will still be accepted just like the previous game_engine_91222.