New game engine!
Features:

.music.stop() has now been added! Instead of the pygame.mixer.music.stop() this can stop the music for you! If an audio driver has not been detected it can prevent errors in the code. Additionally, it reduces the complexity of having to check for an audio driver every time you want to stop the audio from playing.

Changelog:

.music.fade_out() - Now only runs if an audio driver is detected

try / except policies - This has been added to the bottom of the game_engine. It will try to import pygame, if it cannot it will automatically run a script to help out the user to install pygame. Additionally, it will try to find an audio driver. If one has not been found, the .music functions will not be available.
Furthermore, a variable called "activate_music" will be created. This can be directly accessed by your code if you need to check for audio drivers. Use <name of imported file>.activate_music to find the value of this boolen.

Minor improvements to the rendering - This now will find the lengh of the list within the range instead of a seperate variable. Reducing the size of the code.